+++ muikku-school-data-pyramus/src/main/java/fi/otavanopisto/muikku/plugins/schooldatapyramus/PyramusWorkspaceSchoolDataBridge.java
@@ -535,41 +535,4 @@ public class PyramusWorkspaceSchoolDataBridge implements WorkspaceSchoolDataBrid
     return response;
   }
 
-  @Override
-  public List<WorkspaceActivity> listWorkspaceActivities(String studentIdentifier, String workspaceIdentifier, boolean includeTransferCredits) {
-    
-    // Convert identifiers to Pyramus ids
-    
-    Long studentId = identifierMapper.getPyramusStudentId(studentIdentifier);
-    Long courseId = identifierMapper.getPyramusCourseId(workspaceIdentifier);
-    
-    // Make the call
-    
-    CourseActivity[] response = pyramusClient.get(
-        String.format("/students/students/%d/courseActivity?courseIds=%d&includeTransferCredits=%s",
-            studentId,
-            courseId,
-            includeTransferCredits), CourseActivity[].class);
-    
-    // Convert Pyramus CourseActivity to Muikku WorkspaceActivity
-    
-    List<String> curriculumIdentifiers = new ArrayList<>();
-    List<WorkspaceActivity> activities = new ArrayList<>();
-    for (int i = 0; i < response.length; i++) {
-      WorkspaceActivity activity = new WorkspaceActivity();
-      activity.setIdentifier(response[i].getCourseId() == null ? null : response[i].getCourseId().toString());
-      for (Long curriculumId : response[i].getCurriculumIds()) {
-        curriculumIdentifiers.add(identifierMapper.getCurriculumIdentifier(curriculumId).toId());
-      }
-      activity.setName(response[i].getCourseName());
-      activity.setGrade(response[i].getGrade());
-      activity.setPassingGrade(response[i].getPassingGrade());
-      activity.setState(WorkspaceActivityState.valueOf(response[i].getState().toString()));
-      activity.setDate(response[i].getActivityDate());
-      activities.add(activity);
-    }
-    
-    return activities;
-  }
-
 }
